Make a parsers directory
