+++ exe/try
@@ -17,7 +17,7 @@ if ENV['COVERAGE'] || ENV['SIMPLECOV']
     add_group 'Core', 'lib/tryouts.rb'
     add_group 'CLI', 'lib/tryouts/cli'
     add_group 'Formatters', 'lib/tryouts/cli/formatters'
-    add_group 'Parsers', 'lib/tryouts/prism_parser.rb'
+    add_group 'Parsers', 'lib/tryouts/parsers'
     add_group 'Data Structures', ['lib/tryouts/testcase.rb', 'lib/tryouts/testbatch.rb']
     add_group 'Translators', 'lib/tryouts/translators'
     add_group 'Execution', ['lib/tryouts/test_executor.rb', 'lib/tryouts/test_runner.rb', 'lib/tryouts/file_processor.rb']
