+++ lib/tryouts/parsers/prism_parser.rb
@@ -1,13 +1,11 @@
-# Enhanced parser using Prism's inhouse comment extraction capabilities
-# Drop-in replacement for PrismParser that eliminates HEREDOC parsing issues
+# lib/tryouts/parsers/prism_parser.rb
 
 require 'prism'
-require_relative 'test_case'
+require_relative '../test_case'
 
 class Tryouts
-  # Enhanced parser that replaces manual line-by-line parsing with inhouse Prism APIs
-  # while maintaining full compatibility with the original parser's logic structure
-  class EnhancedParser
+  # Fixed PrismParser with pattern matching for robust token filtering
+  class PrismParser
     def initialize(source_path)
       @source_path  = source_path
       @source       = File.read(source_path)
@@ -19,9 +17,7 @@ class Tryouts
     def parse
       return handle_syntax_errors if @prism_result.failure?
 
-      # Use inhouse comment extraction instead of line-by-line regex parsing
-      # This automatically excludes HEREDOC content!
-      tokens           = tokenize_content_with_inhouse_extraction
+      tokens           = tokenize_content
       test_boundaries  = find_test_case_boundaries(tokens)
       tokens           = classify_potential_descriptions_with_boundaries(tokens, test_boundaries)
       test_blocks      = group_into_test_blocks(tokens)
@@ -30,95 +26,66 @@ class Tryouts
 
     private
 
-    # Inhouse comment extraction - replaces the manual regex parsing
-    def tokenize_content_with_inhouse_extraction
+    # Tokenize content using pattern matching for clean line classification
+    def tokenize_content
       tokens = []
 
-      # Get all comments using inhouse Prism extraction
-      comments        = Prism.parse_comments(@source)
-      comment_by_line = comments.group_by { |comment| comment.location.start_line }
-
-      # Process each line, handling multiple comments per line
       @lines.each_with_index do |line, index|
-        line_number = index + 1
-
-        if (comments_for_line = comment_by_line[line_number]) && !comments_for_line.empty?
-          emitted_code = false
-          comments_for_line.sort_by! { |c| c.location.start_column }
-          comments_for_line.each do |comment|
-            comment_content = comment.slice.strip
-            if comment.location.start_column > 0
-              unless emitted_code
-                tokens << { type: :code, content: line, line: index, ast: parse_ruby_line(line) }
-                emitted_code = true
-              end
-              # Inline comment may carry expectations; classify it too
-              tokens << classify_comment_inhousely(comment_content, line_number)
-            else
-              tokens << classify_comment_inhousely(comment_content, line_number)
-            end
-          end
-          next
-        end
-
-        # Handle non-comment lines (blank lines and code)
         token = case line
-                when /^\s*$/
+                in /^##\s*(.*)$/ # Test description format: ## description
+                  { type: :description, content: $1.strip, line: index }
+                in /^#\s*TEST\s*\d*:\s*(.*)$/  # rubocop:disable Lint/DuplicateBranch
+                  { type: :description, content: $1.strip, line: index }
+                in /^#\s*=!>\s*(.*)$/ # Exception expectation (updated for consistency)
+                  { type: :exception_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^#\s*=<>\s*(.*)$/ # Intentional failure expectation
+                  { type: :intentional_failure_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^#\s*==>\s*(.*)$/ # Boolean true expectation
+                  { type: :true_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in %r{^#\s*=/=>\s*(.*)$} # Boolean false expectation
+                  { type: :false_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^#\s*=\|>\s*(.*)$/ # Boolean (true or false) expectation
+                  { type: :boolean_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^#\s*=:>\s*(.*)$/ # Result type expectation
+                  { type: :result_type_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^#\s*=~>\s*(.*)$/ # Regex match expectation
+                  { type: :regex_match_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^#\s*=%>\s*(.*)$/ # Performance time expectation
+                  { type: :performance_time_expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^#\s*=(\d+)>\s*(.*)$/ # Output expectation (stdout/stderr with pipe number)
+                  { type: :output_expectation, content: $2.strip, pipe: $1.to_i, line: index, ast: parse_expectation($2.strip) }
+                in /^#\s*=>\s*(.*)$/ # Regular expectation
+                  { type: :expectation, content: $1.strip, line: index, ast: parse_expectation($1.strip) }
+                in /^##\s*=>\s*(.*)$/ # Commented out expectation (should be ignored)
+                  { type: :comment, content: '=>' + $1.strip, line: index }
+                in /^#\s*(.*)$/ # Single hash comment - potential description
+                  { type: :potential_description, content: $1.strip, line: index }
+                in /^\s*$/ # Blank line
                   { type: :blank, line: index }
-                else
+                else # Ruby code
                   { type: :code, content: line, line: index, ast: parse_ruby_line(line) }
                 end
+
         tokens << token
       end
 
+      # Return tokens with potential_descriptions - they'll be classified later with test boundaries
       tokens
     end
 
-    # Inhouse comment classification - replaces complex regex patterns
-    def classify_comment_inhousely(content, line_number)
-      case content
-      when /^##\s*(.*)$/
-        { type: :description, content: $1.strip, line: line_number - 1 }
-      when /^#\s*TEST\s*\d*:\s*(.*)$/
-        { type: :description, content: $1.strip, line: line_number - 1 }
-      when /^#\s*=!>\s*(.*)$/
-        { type: :exception_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^#\s*=<>\s*(.*)$/
-        { type: :intentional_failure_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^#\s*==>\s*(.*)$/
-        { type: :true_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when %r{^#\s*=/=>\s*(.*)$}
-        { type: :false_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^#\s*=\|>\s*(.*)$/
-        { type: :boolean_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^#\s*=:>\s*(.*)$/
-        { type: :result_type_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^#\s*=~>\s*(.*)$/
-        { type: :regex_match_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^#\s*=%>\s*(.*)$/
-        { type: :performance_time_expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^#\s*=(\d+)>\s*(.*)$/
-        { type: :output_expectation, content: $2.strip, pipe: $1.to_i, line: line_number - 1, ast: parse_expectation($2.strip) }
-      when /^#\s*=>\s*(.*)$/
-        { type: :expectation, content: $1.strip, line: line_number - 1, ast: parse_expectation($1.strip) }
-      when /^##\s*=>\s*(.*)$/
-        { type: :comment, content: '=>' + $1.strip, line: line_number - 1 }
-      when /^#\s*(.*)$/
-        { type: :potential_description, content: $1.strip, line: line_number - 1 }
-      else
-        { type: :comment, content: content.sub(/^#\s*/, ''), line: line_number - 1 }
-      end
-    end
-
-    # Copy the rest of the methods from PrismParser to maintain identical behavior
-
+    # Find actual test case boundaries by looking for ## descriptions or # TEST: patterns
+    # followed by code and expectations
     def find_test_case_boundaries(tokens)
       boundaries = []
 
       tokens.each_with_index do |token, index|
+        # Look for explicit test descriptions (## or # TEST:)
         if token[:type] == :description
+          # Find the end of this test case by looking for the last expectation
+          # before the next description or end of file
           start_line = token[:line]
           end_line   = find_test_case_end(tokens, index)
+
           boundaries << { start: start_line, end: end_line } if end_line
         end
       end
@@ -126,13 +93,19 @@ class Tryouts
       boundaries
     end
 
+    # Find where a test case ends by looking for the last expectation
+    # before the next test description or end of tokens
     def find_test_case_end(tokens, start_index)
       last_expectation_line = nil
 
+      # Look forward from the description for expectations
       (start_index + 1).upto(tokens.length - 1) do |i|
         token = tokens[i]
+
+        # Stop if we hit another test description
         break if token[:type] == :description
 
+        # Track the last expectation we see
         if is_expectation_type?(token[:type])
           last_expectation_line = token[:line]
         end
@@ -141,34 +114,50 @@ class Tryouts
       last_expectation_line
     end
 
+    # Convert potential_descriptions to descriptions or comments using test case boundaries
     def classify_potential_descriptions_with_boundaries(tokens, test_boundaries)
       tokens.map.with_index do |token, index|
         if token[:type] == :potential_description
+          # Check if this comment falls within any test case boundary
           line_num         = token[:line]
           within_test_case = test_boundaries.any? do |boundary|
             line_num >= boundary[:start] && line_num <= boundary[:end]
           end
 
           if within_test_case
+            # This comment is within a test case, treat as regular comment
             token.merge(type: :comment)
           else
+            # For comments outside test boundaries, be more conservative
+            # Only treat as description if it immediately precedes a test pattern AND
+            # looks like a test description
             content = token[:content].strip
 
+            # Check if this looks like a test description based on content
             looks_like_test_description = content.match?(/test|example|demonstrate|show|should|when|given/i) &&
                                           content.length > 10
 
+            # Check if there's code immediately before this (suggesting it's mid-test)
             prev_token      = index > 0 ? tokens[index - 1] : nil
             has_code_before = prev_token && prev_token[:type] == :code
 
             if has_code_before || !looks_like_test_description
+              # Treat as regular comment
               token.merge(type: :comment)
             else
-              following_tokens     = tokens[(index + 1)..]
-              meaningful_following = following_tokens.reject { |t| [:blank, :comment].include?(t[:type]) }
-              test_window          = meaningful_following.first(5)
-              has_code             = test_window.any? { |t| t[:type] == :code }
-              has_expectation      = test_window.any? { |t| is_expectation_type?(t[:type]) }
+              # Look ahead for IMMEDIATE test pattern (stricter than before)
+              following_tokens = tokens[(index + 1)..]
 
+              # Skip blanks and comments to find meaningful content
+              meaningful_following = following_tokens.reject { |t| [:blank, :comment].include?(t[:type]) }
+
+              # Look for test pattern within next 5 tokens (more restrictive)
+              test_window     = meaningful_following.first(5)
+              has_code        = test_window.any? { |t| t[:type] == :code }
+              has_expectation = test_window.any? { |t| is_expectation_type?(t[:type]) }
+
+              # Only promote to description if BOTH code and expectation are found nearby
+              # AND it looks like a test description
               if has_code && has_expectation && looks_like_test_description
                 token.merge(type: :description)
               else
@@ -182,6 +171,52 @@ class Tryouts
       end
     end
 
+    # Convert potential_descriptions to descriptions or comments based on context
+    def classify_potential_descriptions(tokens)
+      tokens.map.with_index do |token, index|
+        if token[:type] == :potential_description
+          # Check if this looks like a test description based on content and context
+          content = token[:content].strip
+
+          # Skip if it's clearly just a regular comment (short, lowercase, etc.)
+          # Test descriptions are typically longer and more descriptive
+          looks_like_regular_comment = content.length < 20 &&
+                                       content.downcase == content &&
+                                       !content.match?(/test|example|demonstrate|show/i)
+
+          # Check if there's code immediately before this (suggesting it's mid-test)
+          prev_token      = index > 0 ? tokens[index - 1] : nil
+          has_code_before = prev_token && prev_token[:type] == :code
+
+          if looks_like_regular_comment || has_code_before
+            # Treat as regular comment
+            token.merge(type: :comment)
+          else
+            # Look ahead for test pattern: code + at least one expectation within reasonable distance
+            following_tokens = tokens[(index + 1)..]
+
+            # Skip blanks and comments to find meaningful content
+            meaningful_following = following_tokens.reject { |t| [:blank, :comment].include?(t[:type]) }
+
+            # Look for test pattern: at least one code token followed by at least one expectation
+            # within the next 10 meaningful tokens (to avoid matching setup/teardown)
+            test_window     = meaningful_following.first(10)
+            has_code        = test_window.any? { |t| t[:type] == :code }
+            has_expectation = test_window.any? { |t| is_expectation_type?(t[:type]) }
+
+            if has_code && has_expectation
+              token.merge(type: :description)
+            else
+              token.merge(type: :comment)
+            end
+          end
+        else
+          token
+        end
+      end
+    end
+
+    # Check if token type represents any kind of expectation
     def is_expectation_type?(type)
       [
         :expectation, :exception_expectation, :intentional_failure_expectation,
@@ -221,6 +256,7 @@ class Tryouts
       has_code && has_expectation
     end
 
+    # Group tokens into logical test blocks using pattern matching
     def group_into_test_blocks(tokens)
       blocks        = []
       current_block = new_test_block
@@ -295,6 +331,7 @@ class Tryouts
       classify_blocks(blocks)
     end
 
+    # Process classified test blocks into domain objects
     def process_test_blocks(classified_blocks)
       setup_blocks    = classified_blocks.filter { |block| block[:type] == :setup }
       test_blocks     = classified_blocks.filter { |block| block[:type] == :test }
@@ -305,7 +342,7 @@ class Tryouts
         test_cases: test_blocks.map { |block| build_test_case(block) },
         teardown: build_teardown(teardown_blocks),
         source_file: @source_path,
-        metadata: { parsed_at: @parsed_at, parser: :enhanced },
+        metadata: { parsed_at: @parsed_at, parser: :prism_v2_fixed },
       )
     end
 
@@ -329,6 +366,8 @@ class Tryouts
       )
     end
 
+    # Modern Ruby 3.4+ pattern matching for robust code extraction
+    # This filters out comments added by add_context_to_block explicitly
     def extract_pure_code_from_blocks(blocks)
       blocks
         .flat_map { |block| block[:code] }
@@ -346,6 +385,7 @@ class Tryouts
     def calculate_block_range(blocks)
       return 0..0 if blocks.empty?
 
+      # Filter out blocks with nil line numbers and build valid ranges
       valid_blocks = blocks.filter { |block| block[:start_line] && block[:end_line] }
       return 0..0 if valid_blocks.empty?
 
@@ -407,12 +447,15 @@ class Tryouts
     def add_context_to_block(block, token)
       case [block[:expectations].empty?, token]
       in [true, { type: :comment | :blank }]
+        # Comments before expectations go with code
         block[:code] << token
       in [false, { type: :comment | :blank }]
+        # Comments after expectations are test context
         block[:comments] << token
       end
     end
 
+    # Classify blocks as setup, test, or teardown based on content
     def classify_blocks(blocks)
       blocks.map.with_index do |block, index|
         block_type = case block
@@ -423,7 +466,7 @@ class Tryouts
                      in { expectations: Array => exps } if !exps.empty?
                        :test
                      else
-                       :preamble
+                       :preamble # Default fallback
                      end
 
         block.merge(type: block_type, end_line: calculate_end_line(block))
@@ -431,6 +474,7 @@ class Tryouts
     end
 
     def calculate_end_line(block)
+      # Only consider actual content (code and expectations), not blank lines/comments
       content_tokens = [*block[:code], *block[:expectations]]
       return block[:start_line] if content_tokens.empty?
 
@@ -447,7 +491,10 @@ class Tryouts
         start_line: Integer => start_line,
         end_line: Integer => end_line
       }
-        source_lines           = @lines[start_line..end_line]
+        # Extract source lines from the original source during parsing
+        source_lines = @lines[start_line..end_line]
+
+        # Find the first expectation line for better error reporting
         first_expectation_line = exp_tokens.empty? ? start_line : exp_tokens.first[:line]
 
         TestCase.new(
@@ -467,6 +514,7 @@ class Tryouts
                    else :regular
                    end
 
+            # For output expectations, we need to preserve the pipe number
             if token[:type] == :output_expectation
               OutputExpectation.new(content: token[:content], type: type, pipe: token[:pipe])
             else
