+++ try/continue-on-error/line_number_edge_cases_debug_noautorun.rb
@@ -2,7 +2,8 @@
 # Debug script for edge cases that might trigger line number off-by-one bugs
 # This tests scenarios that could cause inconsistencies in line number tracking
 
-require_relative '../../lib/tryouts/prism_parser'
+require_relative '../../lib/tryouts/parsers/prism_parser'
+require_relative '../../lib/tryouts/parsers/enhacned_parser'
 
 puts "=== Line Number Edge Cases Debug ==="
 puts
