+++ try/continue-on-error/line_number_real_test_noautorun.rb
@@ -2,7 +2,8 @@
 # Real test to verify line number consistency fix in practice
 # This creates actual failing tests and checks error output for consistent line numbers
 
-require_relative '../../lib/tryouts/prism_parser'
+require_relative '../../lib/tryouts/parsers/prism_parser'
+require_relative '../../lib/tryouts/parsers/enhacned_parser'
 # Don't require verbose formatter to avoid dependency issues
 require_relative '../../lib/tryouts/failure_collector'
 require 'tempfile'
